updated silenceTheshold as a float and parametrized

+++ src/Settings/AudioAnalyzer/PMSettingsManagerAudioAnalyzers.h
@@ -27,7 +27,7 @@ public:
     float getMaxEnergy();
 
     bool getSilenceEnabled();
-    int getSilenceThreshold();
+    float getSilenceThreshold();
     unsigned int getSilenceQueueLength();
 
     bool getMelBandsEnabled();
